Labfiles/04-integrate-agent-with-foundry-iq/Python/requirements.txt
azure-ai-projects==2.0.0b4
azure-identity
python-dotenv
flask
